+++ decorator/TaxComputation.java
@@ -1,6 +1,6 @@
 import java.util.Calendar;
 
 public abstract class TaxComputation {
-    public abstract double computeTax(ShoppingCart shoppingCart, Calendar date);
+    public abstract float computeTax(ShoppingCart shoppingCart, Calendar date);
     protected abstract boolean taxHoliday(Calendar date);
 }
